+++ items.rb
@@ -1,8 +1,8 @@
 class Items
   attr_accessor :publish_date, :label, :genre, :author, :archived
 
-  def initialize(id)
-    @id = id
+  def initialize(_id)
+    @id = Random.rand(1..1000)
     @publish_date = publish_date
     @archived = false
   end
